Move testing of read_write module in own testing module.
